refactor: extract 20 widget sections into individual files
Split UnifiedDashboardView.swift (~2700 lines) into modular widget files
under Views/Dashboard/Widgets/. Each widget section is now a standalone
SwiftUI view with clean init parameters. Shared models and components
extracted to UnifiedWidgetModels.swift and UnifiedWidgetComponents.swift.
